fix: make localstorage defined

+++ utils/cache.ts
@@ -1,3 +1,5 @@
+'use client'
+
 import { LoginResponse } from '@vaika-api/typescript-client';
 
 export interface CacheObject<T> {
@@ -8,14 +10,14 @@ export interface CacheObject<T> {
 }
 
 const createObjectCacher = <T>(key: string, storageFactory: () => Storage) => {
-  const storage = storageFactory();
+  const storage: Storage = storageFactory();
   return {
     replace: (obj: T) => {
-      storage.setItem(key, JSON.stringify(obj));
+      storage?.setItem(key, JSON.stringify(obj));
       return obj;
     },
     get: () => {
-      const obj = storage.getItem(key);
+      const obj = storage?.getItem(key);
       if (!obj) return null;
       try {
         return JSON.parse(obj) as T;
@@ -24,17 +26,24 @@ const createObjectCacher = <T>(key: string, storageFactory: () => Storage) => {
       }
     },
     invalidate: () => {
-      storage.removeItem(key);
+      storage?.removeItem(key);
     },
-    isPresent: () => storage.getItem(key) != null,
+    isPresent: () => storage?.getItem(key) != null,
   };
 };
 
-const inLocalStorage = <T>(key: string) => createObjectCacher<T>(key, () => localStorage);
+// @ts-ignore
+const inLocalStorage = <T>(key: string) => createObjectCacher<T>(key, () => {
+  if (typeof window !== 'undefined') {
+    return window?.localStorage;
+  }
+});
 
 export const authTokenCache = inLocalStorage<LoginResponse>('auth_token');
 
 export const clearCaches = () => {
-  localStorage.clear();
-  sessionStorage.clear();
+  if (typeof window !== 'undefined') {
+    localStorage.clear();
+    sessionStorage.clear();
+  }
 };
